apps/agents/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
anthropic
langgraph
langchain-anthropic
playwright
beautifulsoup4
pillow
pydantic
voyageai
